# docs/swagger.yaml
basePath: /
definitions:
  dnsapi.CreateRecordRequest:
    properties:
      name:
        example: www
        type: string
      prio:
        example: 0
        type: integer
      ttl:
        example: 3600
        type: integer
      type:
        example: A
        type: string
      value:
        example: 1.2.3.4
        type: string
    type: object
  dnsapi.CreateZoneRequest:
    properties:
      abuse_email:
        example: admin@example.com
        type: string
      domain:
        example: example.com
        type: string
      owner:
        example: alice
        type: string
      tags:
        example: tag1,tag2
        type: string
    type: object
...
  dnsapi.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  dnsapi.MessageResponse:
    properties:
      message:
        type: string
    type: object
  dnsapi.Record:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      prio:
        type: integer
      ttl:
        type: integer
      type:
        description: A, AAAA, CNAME, TXT, SRV
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  dnsapi.UpdateRecordRequest:
    properties:
      name:
        example: www
        type: string
      prio:
        example: 0
        type: integer
      ttl:
        example: 3600
        type: integer
      value:
        example: 1.2.3.4
        type: string
    type: object
  dnsapi.UpdateZoneRequest:
    properties:
      abuse_email:
        example: admin@example.com
        type: string
      owner:
        example: alice
        type: string
      tags:
        example: tag1,tag2
        type: string
    type: object
  dnsapi.Zone:
    properties:
      abuse_email:
        type: string
      committed_serial:
        type: string
      created_at:
        type: string
      delete:
        type: boolean
      domain:
        type: string
      id:
        type: integer
      owner:
        type: string
      records:
        items:
          $ref: '#/definitions/dnsapi.Record'
        type: array
      serial:
        type: string
      tags:
        description: Tags separated by comma
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: Simple DNS API for managing Bind zones and records.
  title: DNS API
  version: "1.1"
paths:
...
  /zones/:
    get:
      description: Return all zones including records.
      operationId: get-zones
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dnsapi.Zone'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List zones
      tags:
      - zones
    post:
      consumes:
      - application/json
      description: Create a new zone.
      operationId: create-zone
      parameters:
      - description: Zone payload
        in: body
        name: zone
        required: true
        schema:
          $ref: '#/definitions/dnsapi.CreateZoneRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dnsapi.Zone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create zone
      tags:
      - zones
  /zones/{zone_id}:
    delete:
      description: Delete a zone and all its records.
      operationId: delete-zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete zone
      tags:
      - zones
    get:
      description: Return one zone including records.
      operationId: get-zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.Zone'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get zone
      tags:
      - zones
    put:
      consumes:
      - application/json
      description: Update zone metadata (tags and abuse email only). Use the records
        endpoints to manage records.
      operationId: update-zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      - description: Zone metadata to update
        in: body
        name: zone
        required: true
        schema:
          $ref: '#/definitions/dnsapi.UpdateZoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.Zone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update zone
      tags:
      - zones
  /zones/{zone_id}/commit:
    put:
      description: Write zone changes into DNS servers.
      operationId: commit-zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Commit zone
      tags:
      - zones
  /zones/{zone_id}/records/:
    get:
      description: Return all records for a zone.
      operationId: get-records
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dnsapi.Record'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List records
      tags:
      - records
    post:
      consumes:
      - application/json
      description: Create a new record in a zone. Call commit afterwards to write
        it to bind.
      operationId: create-record
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      - description: Record payload
        in: body
        name: record
        required: true
        schema:
          $ref: '#/definitions/dnsapi.CreateRecordRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dnsapi.Record'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create record
      tags:
      - records
  /zones/{zone_id}/records/{record_id}:
    delete:
      description: Delete a record within a zone.
      operationId: delete-record
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      - description: Record ID
        in: path
        name: record_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete record
      tags:
      - records
    get:
      description: Return one record within a zone.
      operationId: get-record
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      - description: Record ID
        in: path
        name: record_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.Record'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get record
      tags:
      - records
    put:
      consumes:
      - application/json
      description: Update an existing record within a zone. Type cannot be changed;
        delete and recreate instead. Call commit afterwards.
      operationId: update-record
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      - description: Record ID
        in: path
        name: record_id
        required: true
        type: integer
      - description: Record fields to update
        in: body
        name: record
        required: true
        schema:
          $ref: '#/definitions/dnsapi.UpdateRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.Record'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update record
      tags:
      - records
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"